Various test changes, and improvements to project wizard test
