--- tests/minimal/meta/guiOptions.json
@@ -1,1 +1,0 @@
-{}
