--- tests/minimal/meta/sessionStats.log
@@ -1,2 +1,0 @@
-# Start Time         End Time                Novel     Notes
-2020-08-28 11:36:36  2020-08-28 11:36:48        10         0
